+++ components/forms/Form/Form.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { cleanup, render, screen, act, waitFor } from "@testing-library/react";
+import { cleanup, render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import Form from "./Form";
 import { submitToAPI } from "@lib/helpers";
@@ -112,41 +112,36 @@ describe("Form Functionality", () => {
   });
 
   it("there is 1 and only 1 submit button", async () => {
-    act(() => {
-      render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
-    });
+    render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
     expect(await screen.findAllByRole("button", { type: "submit" })).toHaveLength(1);
   });
 
   it("Form can be submitted", async () => {
-    userEvent.setup();
-    act(() => {
-      render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
-    });
+    const user = userEvent.setup();
+    render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
     expect(screen.getByRole("button", { type: "submit" })).toBeInTheDocument();
 
-    await act(async () => await userEvent.click(screen.getByRole("button", { type: "submit" })));
+    await user.click(screen.getByRole("button", { type: "submit" }));
 
     await waitFor(() => expect(submitToAPI).toBeCalledTimes(1));
   });
 
   it("shows the alert after pressing submit if the timer hasn't expired", async () => {
-    userEvent.setup();
+    const user = userEvent.setup();
     mockFormTimerState = {
       canSubmit: false,
       remainingTime: 5,
       timerDelay: 5,
       timeLock: 0,
     };
-    act(() => {
-      render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
-    });
+
+    render(<Form formRecord={formRecord} language="en" t={(key) => key} />);
     const submitButton = screen.getByRole("button", { type: "submit" });
     await waitFor(() => expect(submitButton).toBeInTheDocument());
 
     expect(screen.queryByRole("alert")).not.toBeInTheDocument();
 
-    await act(async () => await userEvent.click(screen.getByRole("button", { type: "submit" })));
+    await user.click(screen.getByRole("button", { type: "submit" }));
     expect(await screen.findByRole("alert")).toBeInTheDocument();
     expect(submitToAPI).not.toBeCalled();
   });