+++ CHANGELOG.md
@@ -7,11 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Validation of a JSON Config to check the IDs of elements [#892](https://github.com/cds-snc/platform-forms-client/pull/892)
+- Added login page [#867](https://github.com/cds-snc/platform-forms-client/issues/867)
+- Added login page for temporary token [#900](https://github.com/cds-snc/platform-forms-client/pull/900)
+
+### Fixed
+
+- Fix stuck "Loading..." animation after uploading a new JSON config. [#898](https://github.com/cds-snc/platform-forms-client/pull/898)
+
 ## [1.3.0] 2022-07-15
 
 ### Added
 
 - Make GC Branding in Footer configurable [#847](https://github.com/cds-snc/platform-forms-client/pull/847)
+- Logout Page [#847] (https://github.com/cds-snc/platform-forms-client/issues/870)
 
 ### Fixed